change type gender user

+++ db/migrate/20150212033349_add_gender_to_users.rb
@@ -1,5 +1,5 @@
 class AddGenderToUsers < ActiveRecord::Migration
   def change
-    add_column :users, :gender, :TINYINT
+    add_column :users, :gender, :integer
   end
 end
